renamed histroy.ts to history-meal.ts
